Перенос бизнес-логики в отдельную папку
