+++ Pavel/CircuitCalculator.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2020
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitCalculator", "CircuitCalculator\CircuitCalculator.csproj", "{8F14EB09-44DC-460B-BEA9-CCCB65F812D4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitModel", "CircuitModel\CircuitModel.csproj", "{8F14EB09-44DC-460B-BEA9-CCCB65F812D4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
